+++ libs/apod-react/src/lib/hooks/index.ts
@@ -1,3 +1,1 @@
-export * from './apod-api-base-url';
 export * from './use-apod';
-
