+++ libs/apod-react/src/lib/utils/get-with-start-and-end-dates.ts
@@ -1,8 +1,14 @@
 import {
   ApodResponse,
-  APOD_API_URL,
   GetWithStartAndEndDatesParams,
 } from '@chrome-neo-plus/apod-common';
+import { httpsCallable } from 'firebase/functions';
+import { functions } from './firebase';
+
+const getApod = httpsCallable<GetWithStartAndEndDatesParams, ApodResponse[]>(
+  functions,
+  'getApod'
+);
 
 /**
  * Returns the APOD data for the given date-range.
@@ -13,11 +19,5 @@ import {
 export async function getWithStartAndEndDates(
   params: GetWithStartAndEndDatesParams
 ): Promise<ApodResponse[]> {
-  const url = new URL(params.baseUrl ?? APOD_API_URL);
-  if (params.api_key) url.searchParams.append('api_key', params.api_key);
-  url.searchParams.append('start_date', params.start_date);
-  if (params.end_date) url.searchParams.append('end_date', params.end_date);
-  if (params.thumbs) url.searchParams.append('thumbs', 'true');
-  const res = await fetch(url.toString());
-  return res.json();
+  return getApod(params).then(({ data }) => data);
 }
